dddddd

+++ FSKit/Classes/FSKit.h
@@ -124,8 +124,8 @@
 + (CGFloat)DEBJWithYearRate:(CGFloat)rate monthes:(NSInteger)month;
 + (CGFloat)DEBXWithYearRate:(CGFloat)rate monthes:(NSInteger)month;
 + (CGFloat)freeStoragePercentage;   // 可用内存占总内存的比例,eg  0.1;
-+ (long long)getTotalDiskSize;   // 获取磁盘总量
-+ (long long)getAvailableDiskSize;   // 获取磁盘可用量
++ (NSInteger)getTotalDiskSize;   // 获取磁盘总量
++ (NSInteger)getAvailableDiskSize;   // 获取磁盘可用量
 
 + (NSString *)appVersionNumber;                                                     // 获得版本号
 + (NSString *)appName;                                                              // 获得应用的Bundle Display Name
